Script for creating and saving plots
Experiments to create and save plots, imiting the real package function.
It gets 3 args, the first variable name, the second variable name and the save path.
If the second variable in not defined, then the plot is created as a 1D plot of the first variable, if they are both defined then the plot is 2D